Create modules by layer
