+++ app/build.gradle
@@ -42,6 +42,10 @@ android {
 
 dependencies {
 
+    implementation project(':domain')
+    implementation project(':data')
+    implementation project(':usecases')
+
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.4.1'
     implementation 'com.google.android.material:material:1.5.0'
@@ -65,6 +69,8 @@ dependencies {
     implementation 'androidx.room:room-ktx:2.6.1'
     kapt 'androidx.room:room-compiler:2.6.1'
     implementation "androidx.room:room-testing:2.6.1"
+
+    implementation "io.arrow-kt:arrow-core:1.0.1"
 
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
