--- app/src/main/java/com/jorgesm/themoviedb/data/PermissionChecker.kt
@@ -1,12 +1,0 @@
-package com.jorgesm.themoviedb.data
-
-import android.app.Application
-import android.content.pm.PackageManager
-import androidx.core.content.ContextCompat
-
-class PermissionChecker(private val application: Application, private val permission: String) {
-    
-    fun check(): Boolean = ContextCompat.checkSelfPermission( application, permission
-    ) == PackageManager.PERMISSION_GRANTED
-    
-}
