--- app/src/main/java/com/jorgesm/themoviedb/data/RegionRepository.kt
@@ -1,36 +1,0 @@
-package com.jorgesm.themoviedb.data
-
-
-import android.Manifest
-import android.app.Application
-import android.location.Geocoder
-import android.location.Location
-import com.jorgesm.themoviedb.data.datasource.LocationDataSource
-import com.jorgesm.themoviedb.framework.PlayServicesLocationDataSource
-
-
-class RegionRepository(application: Application) {
-    companion object{
-        private const val DEFAULT_REGION = "US"
-    }
-
-    private val locationDataSource: LocationDataSource = PlayServicesLocationDataSource(application)
-    private val coarsePermissionChecker = PermissionChecker(application, Manifest.permission.ACCESS_COARSE_LOCATION)
-    
-    private val geocoder = Geocoder(application)
-    
-    suspend fun findLasRegion(): String = findLastLocation().toRegion()
-    
-    private suspend fun findLastLocation(): Location? {
-        val success = coarsePermissionChecker.check()
-        return if(success) locationDataSource.findLastLocation() else null
-    }
-    
-    
-    private fun Location?.toRegion(): String{
-        val addresses = this?.let {
-            geocoder.getFromLocation(latitude, longitude, 1)
-        }
-        return addresses?.firstOrNull()?.countryName ?: DEFAULT_REGION
-    }
-}
