--- app/src/main/java/com/jorgesm/themoviedb/data/datasource/MovieLocalDataSource.kt
@@ -1,13 +1,0 @@
-package com.jorgesm.themoviedb.data.datasource
-
-import com.jorgesm.themoviedb.domain.DomainMovie
-import kotlinx.coroutines.flow.Flow
-
-interface MovieLocalDataSource {
-    val movies: Flow<List<DomainMovie>>
-    
-    suspend fun isEmpty(): Boolean
-    fun findMovieById(id: Int): Flow<DomainMovie>
-    
-    suspend fun save(movies: List<DomainMovie>)
-}
