--- app/src/main/java/com/jorgesm/themoviedb/data/datasource/MovieRemoteDataSource.kt
@@ -1,8 +1,0 @@
-package com.jorgesm.themoviedb.data.datasource
-
-import com.jorgesm.themoviedb.domain.DomainMovie
-
-interface MovieRemoteDataSource{
-    suspend fun findPopularMovies( region: String): List<DomainMovie>
-    
-}
