+++ app/src/main/java/com/jorgesm/themoviedb/data/server/AndroidPermissionChecker.kt
@@ -1,0 +1,21 @@
+package com.jorgesm.themoviedb.data.server
+
+import android.Manifest
+import android.app.Application
+import android.content.pm.PackageManager
+import androidx.core.content.ContextCompat
+import com.jorgesm.themoviedb.data.PermissionChecker
+
+class AndroidPermissionChecker(private val application: Application) :
+    PermissionChecker {
+    
+    override fun check(permission: PermissionChecker.Permission): Boolean =
+        ContextCompat.checkSelfPermission(
+            application,
+            permission.toAndroidId()
+        ) == PackageManager.PERMISSION_GRANTED
+}
+
+private fun PermissionChecker.Permission.toAndroidId() = when (this) {
+    PermissionChecker.Permission.COARSE_LOCATION -> Manifest.permission.ACCESS_COARSE_LOCATION
+}
