+++ app/src/main/java/com/jorgesm/themoviedb/data/server/RemoteConnection.kt
@@ -1,0 +1,27 @@
+package com.jorgesm.themoviedb.data.server
+
+import com.jorgesm.themoviedb.utils.Constants
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import retrofit2.create
+
+
+object RemoteConnection {
+
+    private val okHttpClient = HttpLoggingInterceptor().run {
+        level = HttpLoggingInterceptor.Level.BODY
+        OkHttpClient.Builder().addInterceptor(this).build()
+    }
+
+    private val builder = Retrofit.Builder()
+        .baseUrl(Constants.BASE_URL)
+        .client(okHttpClient)
+        .addConverterFactory(GsonConverterFactory.create())
+        .build()
+
+    val service: RemoteService = builder.create()
+}
+
+
