+++ app/src/main/java/com/jorgesm/themoviedb/data/server/extensions.kt
@@ -1,0 +1,21 @@
+package com.jorgesm.themoviedb.data.server
+
+import arrow.core.Either
+import arrow.core.left
+import arrow.core.right
+import retrofit2.HttpException
+import java.io.IOException
+import com.jorgesm.themoviedb.domain.Error
+
+
+fun Throwable.toError(): Error = when (this) {
+    is IOException -> Error.Connectivity
+    is HttpException -> Error.Server(code())
+    else -> Error.Unknown(message ?: "")
+}
+
+suspend fun <T> tryCall(action: suspend () -> T): Either<Error, T> = try {
+    action().right()
+} catch (e: Exception) {
+    e.toError().left()
+}
