--- app/src/main/java/com/jorgesm/themoviedb/domain/Error.kt
@@ -1,23 +1,0 @@
-package com.jorgesm.themoviedb.domain
-
-import okio.IOException
-import retrofit2.HttpException
-
-sealed interface Error {
-    class Server(val code: Int): Error
-    object Connectivity: Error
-    class Unknown(val  message: String): Error
-}
-
-fun Throwable.toError(): Error = when (this){
-    is IOException -> Error.Connectivity
-    is HttpException -> Error.Server(code())
-    else -> Error.Unknown(message ?: "")
-}
-
-inline fun <T> tryCall(action: () -> T): Error? = try {
-    action()
-    null
-}catch (e: Exception){
-    e.toError()
-}
