+++ app/src/main/java/com/jorgesm/themoviedb/domain/Mappers.kt
@@ -1,21 +1,22 @@
 package com.jorgesm.themoviedb.domain
 
-import com.jorgesm.themoviedb.framework.database.Movie
+import com.jorgesm.themoviedb.data.database.Movie
 
 fun List<Movie>.toDomainModel(): List<DomainMovie> = map { it.toDomainModel() }
-fun Movie.toDomainModel(): DomainMovie = DomainMovie (
-   id = this.id,
-   title = this.title,
-   releaseDate = this.releaseDate,
-   posterPath = this.posterPath,
-   backdropPath = this.backdropPath ?: "",
-   originalLanguage =this.originalLanguage,
-   originalTitle =this.originalTitle,
-   overview = this.overview,
-   popularity =this.popularity,
-   voteAverage = this.voteAverage,
-   isFavorite = this.isFavorite
-)
+fun Movie.toDomainModel(): DomainMovie =
+   DomainMovie(
+      id = this.id,
+      title = this.title,
+      releaseDate = this.releaseDate,
+      posterPath = this.posterPath,
+      backdropPath = this.backdropPath ?: "",
+      originalLanguage = this.originalLanguage,
+      originalTitle = this.originalTitle,
+      overview = this.overview,
+      popularity = this.popularity,
+      voteAverage = this.voteAverage,
+      isFavorite = this.isFavorite
+   )
 fun List<DomainMovie>.toDBModel(): List<Movie> = map { it.toDBModel() }
 
 fun DomainMovie.toDBModel(): Movie = Movie(
