--- app/src/main/java/com/jorgesm/themoviedb/framework/database/MovieDataBase.kt
@@ -1,9 +1,0 @@
-package com.jorgesm.themoviedb.framework.database
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-
-@Database(entities = [Movie::class], version = 1, exportSchema = false)
-abstract class MovieDataBase: RoomDatabase() {
-    abstract fun movieDao(): MovieDao
-}
