--- app/src/main/java/com/jorgesm/themoviedb/framework/server/RemoteResult.kt
@@ -1,41 +1,0 @@
-package com.jorgesm.themoviedb.framework.server
-
-
-import com.google.gson.annotations.SerializedName
-import java.io.Serializable
-
-
-data class RemoteResult(
-    val page: Int,
-    val results: List<RemoteMovie>,
-    @SerializedName("total_pages")
-    val totalPages: Int,
-    @SerializedName("total_results")
-    val totalResults: Int
-)
-
-
-data class RemoteMovie(
-    val adult: Boolean,
-    @SerializedName("backdrop_path")
-    val backdropPath: String?,
-    @SerializedName("genre_ids")
-    val genreIds: List<Int>,
-    val id: Int,
-    @SerializedName("original_language")
-    val originalLanguage: String,
-    @SerializedName("original_title")
-    val originalTitle: String,
-    val overview: String,
-    val popularity: Double,
-    @SerializedName("poster_path")
-    val posterPath: String,
-    @SerializedName("release_date")
-    val releaseDate: String,
-    val title: String,
-    val video: Boolean,
-    @SerializedName("vote_average")
-    val voteAverage: Double,
-    @SerializedName("vote_count")
-    val voteCount: Int
-): Serializable
