--- app/src/main/java/com/jorgesm/themoviedb/usecases/GetMovieByIdUseCase.kt
@@ -1,8 +1,0 @@
-package com.jorgesm.themoviedb.usecases
-
-import com.jorgesm.themoviedb.data.MoviesRepository
-
-class GetMovieByIdUseCase(private val repository: MoviesRepository) {
-    
-    operator fun invoke(id:Int) = repository.findMovieById(id)
-}
