+++ data/src/main/java/com/jorgesm/themoviedb/data/PermissionChecker.kt
@@ -1,0 +1,9 @@
+package com.jorgesm.themoviedb.data
+
+
+interface PermissionChecker {
+    
+    enum class Permission{ COARSE_LOCATION}
+    fun check(permission: Permission): Boolean
+    
+}
