+++ data/src/main/java/com/jorgesm/themoviedb/data/datasource/LocationDataSource.kt
@@ -1,0 +1,8 @@
+package com.jorgesm.themoviedb.data.datasource
+
+
+
+interface LocationDataSource {
+    suspend fun findLastLocation(): String?
+}
+
