+++ domain/src/main/java/com/jorgesm/themoviedb/domain/Error.kt
@@ -1,0 +1,11 @@
+package com.jorgesm.themoviedb.domain
+
+
+
+
+sealed interface Error {
+    class Server(val code: Int): Error
+    object Connectivity: Error
+    class Unknown(val  message: String): Error
+}
+
