+++ settings.gradle
@@ -14,3 +14,6 @@ dependencyResolutionManagement {
 }
 rootProject.name = "the movie db"
 include ':app'
+include ':domain'
+include ':data'
+include ':usecases'
