+++ usecases/src/main/java/com/jorgesm/themoviedb/usecases/GetPopularMoviesUseCase.kt
@@ -1,0 +1,7 @@
+package com.jorgesm.themoviedb.usecases
+
+import com.jorgesm.themoviedb.data.MoviesRepository
+
+class GetPopularMoviesUseCase(private val repository: MoviesRepository) {
+    operator fun invoke() = repository.popularMovies
+}
