+++ usecases/src/main/java/com/jorgesm/themoviedb/usecases/RequestPopularMoviesUseCase.kt
@@ -1,0 +1,12 @@
+package com.jorgesm.themoviedb.usecases
+
+import com.jorgesm.themoviedb.data.MoviesRepository
+import com.jorgesm.themoviedb.domain.Error
+
+class RequestPopularMoviesUseCase(private val moviesRepository: MoviesRepository) {
+        suspend operator fun invoke(): Error? {
+                return moviesRepository.requestPopularMovies()
+        }
+        
+}
+
